reorganized files into components folder
